+++ src/Util.js
@@ -2,15 +2,16 @@ const Constants = require('./Constants');
 
 var request = require("request");
 
-const callHttpApi = function(log, urlToCall, urlMethod, urlBody, urlForm, urlHeaders, homeKitCallback, context, onSuccessCallback, onFailureCallback, timeout) {
+const callHttpApi = function(log, urlToCall, urlMethod, urlBody, urlForm, urlHeaders, rejectUnauthorized, homeKitCallback, context, onSuccessCallback, onFailureCallback, timeout) {
   if (urlToCall !== "" && context !== Constants.CONTEXT_FROM_WEBHOOK) {
     var theRequest = {
       method : urlMethod,
       url : urlToCall,
       timeout : timeout || Constants.DEFAULT_REQUEST_TIMEOUT,
-      headers : JSON.parse(urlHeaders)
+      headers : JSON.parse(urlHeaders),
+      rejectUnauthorized: rejectUnauthorized
     };
-    if (urlMethod === "POST" || urlMethod === "PUT") {
+    if (urlMethod === "POST" || urlMethod === "PUT" || urlMethod === "PATCH") {
       if (urlForm) {
         log("Adding Form " + urlForm);
         theRequest.form = JSON.parse(urlForm);