+++ src/homekit/accessories/HttpWebHookLockMechanismAccessory.js
@@ -12,6 +12,7 @@ function HttpWebHookLockMechanismAccessory(ServiceParam, CharacteristicParam, pl
   this.id = lockMechanismOpenerConfig["id"];
   this.name = lockMechanismOpenerConfig["name"];
   this.type = "lockmechanism";
+  this.rejectUnauthorized = lockMechanismOpenerConfig["rejectUnauthorized"] === undefined ? true: lockMechanismOpenerConfig["rejectUnauthorized"] === true;
   this.setLockTargetStateOpenURL = lockMechanismOpenerConfig["open_url"] || "";
   this.setLockTargetStateOpenMethod = lockMechanismOpenerConfig["open_method"] || "GET";
   this.setLockTargetStateOpenBody = lockMechanismOpenerConfig["open_body"] || "";
@@ -35,7 +36,7 @@ function HttpWebHookLockMechanismAccessory(ServiceParam, CharacteristicParam, pl
 
 HttpWebHookLockMechanismAccessory.prototype.changeFromServer = function(urlParams) {
   if (urlParams.lockcurrentstate != null) {
-    var cachedLockCurrentState = this.storage.getItemSync("http-webhook--lock-current-state-" + this.id);
+    var cachedLockCurrentState = this.storage.getItemSync("http-webhook-lock-current-state-" + this.id);
     if (cachedLockCurrentState === undefined) {
       cachedLockCurrentState = Characteristic.LockCurrentState.SECURED;
     }
@@ -68,7 +69,7 @@ HttpWebHookLockMechanismAccessory.prototype.changeFromServer = function(urlParam
 }
 
 HttpWebHookLockMechanismAccessory.prototype.getLockTargetState = function(callback) {
-  this.log("Getting current Target Lock State for '%s'...", this.id);
+  this.log.debug("Getting current Target Lock State for '%s'...", this.id);
   var state = this.storage.getItemSync("http-webhook-lock-target-state-" + this.id);
   if (state === undefined) {
     state = Characteristic.LockTargetState.SECURED;
@@ -97,7 +98,7 @@ HttpWebHookLockMechanismAccessory.prototype.setLockTargetState = function(homeKi
     urlHeaders = this.setLockTargetStateOpenHeaders;
   }
 
-  Util.callHttpApi(this.log, urlToCall, urlMethod, urlBody, urlForm, urlHeaders, callback, context, (function() {
+  Util.callHttpApi(this.log, urlToCall, urlMethod, urlBody, urlForm, urlHeaders, this.rejectUnauthorized, callback, context, (function() {
     this.storage.setItemSync("http-webhook-lock-current-state-" + this.id, newHomeKitState);
     this.service.getCharacteristic(Characteristic.LockTargetState).updateValue(newHomeKitStateTarget, undefined, null);
     this.service.getCharacteristic(Characteristic.LockCurrentState).updateValue(newHomeKitState, undefined, null);
@@ -109,7 +110,7 @@ HttpWebHookLockMechanismAccessory.prototype.setLockTargetState = function(homeKi
 };
 
 HttpWebHookLockMechanismAccessory.prototype.getLockCurrentState = function(callback) {
-  this.log("Getting Current Lock State for '%s'...", this.id);
+  this.log.debug("Getting Current Lock State for '%s'...", this.id);
   var state = this.storage.getItemSync("http-webhook-lock-current-state-" + this.id);
   if (state === undefined) {
     state = Characteristic.LockCurrentState.SECURED;