+++ src/homekit/accessories/HttpWebHookSwitchAccessory.js
@@ -11,6 +11,7 @@ function HttpWebHookSwitchAccessory(ServiceParam, CharacteristicParam, platform,
 
   this.id = switchConfig["id"];
   this.name = switchConfig["name"];
+  this.rejectUnauthorized = switchConfig["rejectUnauthorized"] === undefined ? true: switchConfig["rejectUnauthorized"] === true;
   this.onURL = switchConfig["on_url"] || "";
   this.onMethod = switchConfig["on_method"] || "GET";
   this.onBody = switchConfig["on_body"] || "";
@@ -59,7 +60,7 @@ HttpWebHookSwitchAccessory.prototype.changeFromServer = function(urlParams) {
 }
 
 HttpWebHookSwitchAccessory.prototype.getState = function(callback) {
-  this.log("Getting current state for '%s'...", this.id);
+  this.log.debug("Getting current state for '%s'...", this.id);
   var state = this.storage.getItemSync("http-webhook-" + this.id);
   if (state === undefined) {
     state = false;
@@ -84,7 +85,7 @@ HttpWebHookSwitchAccessory.prototype.setState = function(powerOn, callback, cont
     urlHeaders = this.offHeaders;
   }
 
-  Util.callHttpApi(this.log, urlToCall, urlMethod, urlBody, urlForm, urlHeaders, callback, context);
+  Util.callHttpApi(this.log, urlToCall, urlMethod, urlBody, urlForm, urlHeaders, this.rejectUnauthorized, callback, context);
 };
 
 HttpWebHookSwitchAccessory.prototype.getServices = function() {